simple deduplication is now live
Went live with a very basic version of deduplication. If two results have the same formatted name and are withing five kilometers of each other only the first (more accurate) result is returned. This behaviour can be suppressed if the optional `no_dedupe` param is supplied. 

Still need to write a blog post explaining this early next week